Small adjustment regarding use of seq() during knot construction
For the knot extension, we use seq(). However, this might lead to the boundary knots of the interior knots not exactly matching the domain boundaries anymore due to numerical inpreciseness of seq(). Thus, we manually assign the correct values for these two knots again to not having problems when using splines::splineDesign later.