Updated clr and densreg to suppress integral warnings
In clr(), the approximated integral is compared to the theoretical value 1 (inverse = FALSE) or 0 (inverse = TRUE) and a warning is returned in case of deviation. I added an argument tolerance to clr() to specify the allowed tolerance for this deviation without giving a warning. It can also be set to FALSE, in which case the integral is not compared to the theoretical value. This is specified in the calls of clr(..., inverse = FALSE) in densreg.R, since the md-smoother ensures the integrate-to-zero constraint and deviations are only due to approximation (e.g. small bin width, etc).